Il bug della colonna permane, sorto e risolto il problema con il database: ogni installazione aveva bisogno del rinnovo della versione passato al Helper per ottenere nuovo db, bisognava quindi incrementare la versione. La UI è quasi terminata. E' stato un errore sviluppare tutto parallelamente, insegnamento per il futuro.

+++ app/src/main/java/customView/CircularIndicatorAnimation.java
@@ -13,7 +13,7 @@ public class CircularIndicatorAnimation extends Animation {
     private float oldAngle;
     private float newAngle;
 
-    public CircularIndicatorAnimation(CircularIndicator circle, int newAngle) {
+    public CircularIndicatorAnimation(CircularIndicator circle, float newAngle) {
         this.oldAngle = circle.getAngle();
         this.newAngle = newAngle;
         this.circle = circle;
